broke up analysis classes
