+++ autofit/non_linear/analysis/__init__.py
@@ -1,1 +1,3 @@
 from .analysis import Analysis
+from .combined import CombinedAnalysis
+from .free_parameter import FreeParameterAnalysis
